data: Move undo worktree capture
The undo data layer keeps checkpoint orchestration, worktree capture, and
restore mechanics in one module.

The module does not provide a boundary for collecting changed worktree
paths, file modes, and before-image blobs, so checkpoint storage has to
carry filesystem probing.

This commit begins separating undo worktree handling by moving changed-path
discovery, mode detection, blob creation, and path snapshot construction
into data.undo_worktree.

The next commit will add tests that keep before-image capture out of undo
orchestration.
